refactor(run): simplify run.py
